WIP for supporting bigints.

+++ src/jp/ac/kobe_u/cs/sugar/csp/PowerLiteral.java
@@ -1,5 +1,6 @@
 package jp.ac.kobe_u.cs.sugar.csp;
 
+import java.math.BigInteger;
 import java.util.Set;
 
 import jp.ac.kobe_u.cs.sugar.SugarException;
@@ -19,7 +20,7 @@ public class PowerLiteral extends ArithmeticLiteral {
 	}
 
 	@Override
-	public long[] getBound(IntegerVariable v) throws SugarException {
+	public BigInteger[] getBound(IntegerVariable v) throws SugarException {
 		return null;
 	}
